+++ footer.php
@@ -13,7 +13,7 @@
 
 	<?php do_action( 'storefront_before_footer' ); ?>
 
-	<footer id="colophon" class="site-footer" role="content-info">
+	<footer id="colophon" class="site-footer" role="contentinfo">
 		<div class="col-full">
 
 			<?php