+++ header.php
@@ -19,9 +19,6 @@
 
 <body <?php body_class(); ?>>
 <div id="page" class="hfeed site">
-	<a class="skip-link screen-reader-text" href="#site-navigation"><?php _e( 'Skip to navigation', 'storefront' ); ?></a>
-	<a class="skip-link screen-reader-text" href="#content"><?php _e( 'Skip to content', 'storefront' ); ?></a>
-
 	<?php
 	do_action( 'storefront_before_header' ); ?>
 
@@ -30,6 +27,7 @@
 
 			<?php
 			/**
+			 * @hooked storefront_skip_links - 0
 			 * @hooked storefront_social_icons - 10
 			 * @hooked storefront_site_branding - 20
 			 * @hooked storefront_secondary_navigation - 30