+++ inc/structure/header.php
@@ -11,13 +11,15 @@ if ( ! function_exists( 'storefront_header_widget_region' ) ) {
 	 * @since  1.0.0
 	 */
 	function storefront_header_widget_region() {
+		if ( is_active_sidebar( 'header-1' ) ) {
 		?>
-		<div class="header-widget-region">
+		<div class="header-widget-region" role="complementary">
 			<div class="col-full">
 				<?php dynamic_sidebar( 'header-1' ); ?>
 			</div>
 		</div>
 		<?php
+		}
 	}
 }
 
@@ -91,3 +93,19 @@ if ( ! function_exists( 'storefront_secondary_navigation' ) ) {
 		<?php
 	}
 }
+
+if ( ! function_exists( 'storefront_skip_links' ) ) {
+	/**
+	 * Skip links
+	 * @since  1.4.1
+	 * @return void
+	 */
+	function storefront_skip_links() {
+		?>
+		<a class="skip-link screen-reader-text" href="#site-navigation"><?php _e( 'Skip to navigation', 'storefront' ); ?></a>
+		<a class="skip-link screen-reader-text" href="#content"><?php _e( 'Skip to content', 'storefront' ); ?></a>
+		<?php
+	}
+}
+
+